+++ crashmail/memmessage.h
@@ -8,22 +8,22 @@
 struct TextChunk
 {
    struct TextChunk *Next;
-   ulong Length;
+   uint32_t Length;
    uchar Data[PKT_CHUNKLEN];
 };
 
 struct Nodes2D
 {
    struct Nodes2D *Next;
-   ushort Nodes;
-   ushort Net[PKT_NUM2D];
-   ushort Node[PKT_NUM2D];
+   uint16_t Nodes;
+   uint16_t Net[PKT_NUM2D];
+   uint16_t Node[PKT_NUM2D];
 };
 
 struct Path
 {
    struct Path *Next;
-   ushort Paths;
+   uint16_t Paths;
    uchar Path[PKT_NUMPATH][100];
 };
 
@@ -37,7 +37,7 @@ struct Path
 
 struct MemMessage
 {
-   ulong msgnum;
+   uint32_t msgnum;
 
    struct Node4D OrigNode;
    struct Node4D DestNode;
@@ -56,22 +56,22 @@ struct MemMessage
    uchar MSGID[80];
    uchar REPLY[80];
 
-   ushort Attr;
-   ushort Cost;
+   uint16_t Attr;
+   uint16_t Cost;
 
    uchar Type;
-   ushort Flags;
+   uint16_t Flags;
    
    struct jbList TextChunks;
    struct jbList SeenBy;
    struct jbList Path;
 };
 
-bool mmAddNodes2DList(struct jbList *list,ushort net,ushort node);
-void mmRemNodes2DList(struct jbList *list,ushort net,ushort node);
+bool mmAddNodes2DList(struct jbList *list,uint16_t net,uint16_t node);
+void mmRemNodes2DList(struct jbList *list,uint16_t net,uint16_t node);
 void mmRemNodes2DListPat(struct jbList *list,struct Node2DPat *pat);
 bool mmAddPath(uchar *str,struct jbList *list);
-bool mmAddBuf(struct jbList *chunklist,uchar *buf,ulong len);
+bool mmAddBuf(struct jbList *chunklist,uchar *buf,uint32_t len);
 bool mmAddLine(struct MemMessage *mm,uchar *buf);
 struct MemMessage *mmAlloc(void);
 void mmClear(struct MemMessage *mm);
